splitting files to hold single classes, fixing imports to fit PEP 8
