+++ src/gamemode.py
@@ -1,4 +1,5 @@
 import pygame
+
 import shared
 
 class GameMode(object):
