+++ src/main.py
@@ -2,7 +2,7 @@
 
 import sys
 
-from game import *
+from game import Game
 
 def main():
     game = Game()
