+++ src/personality.py
@@ -28,21 +28,3 @@ class Personality(object):
         else:# cls.Aggressive, cls.Energetic
             temp_name = random.choice(("Ga","Ku","Zi","Ru","Te","The")) + random.choice(("va","iy","na","ran"))
         return temp_name + random.choice(("ex","ul","av","em","ix","ab","ev","og","za","el"))
-
-class Mood(object):
-    # So instead of using numbers/strings/whatever for moods, just use Mood.the_mood
-    # This way, if something is mistyped or whatever we will get an error, useful.
-    class Neutral(object):
-        drvChange = 0
-
-    class Bored(object):
-        drvChange = -1
-
-    class Sad(object):
-        drvChange = -1
-
-    class Angry(object):
-        drvChange = 1
-
-    class Happy(object):
-        drvChange = 1
